Time: 11 ms (42.74%), Space: 161.2 MB (28.82%) - LeetHub

+++ 3788-maximum-score-of-a-split/3788-maximum-score-of-a-split.cpp
@@ -17,7 +17,7 @@ public:
 
         long long ans = LLONG_MIN;
         for(int i=0; i<n-1; i++) {
-            ans = max(ans , (long long)(prefix[i] - suffix[i+1]));
+            ans = max(ans , prefix[i] - suffix[i+1]);
         }
         return ans;
     }
